Adicionando NavBarProjectsProvider em index do projeto FinHelper
